hook refactoring; posts now displaying
