+++ src/Kambaz/types.tsx
@@ -1,24 +1,24 @@
 export interface Post {
     _id?: string; // MongoDB ObjectId stored as a string
-    folderId: string; 
-    authorId: string; 
-    datePosted: string; 
+    folderId: string;
+    authorId: string;
+    datePosted: string;
     type: number;
     instructor: boolean;
     title: string;
     content: string;
-    followupDiscussions: string[]; 
-    studentAnswer: string | null; 
-    instructorAnswer: string | null; 
-    viewers: string[]; 
-    courseId: string; 
+    followupDiscussions: string[];
+    studentAnswer: string | null;
+    instructorAnswer: string | null;
+    viewers: string[];
+    courseId: string;
 }
 
 export interface User {
     _id?: string;
     name: string;
     email: string;
-    profilePhoto?: string; 
+    profilePhoto?: string;
     type: number; // 0 = student, 1 = instructor
 }
 
@@ -26,40 +26,40 @@ export interface Course {
     _id?: string;
     name: string;
     term: string;
-    number: string; 
-    students: string[]; 
-    instructors: string[]; 
-    folders: string[]; 
+    number: string;
+    students: string[];
+    instructors: string[];
+    folders: string[];
 }
 
 export interface Folder {
     _id?: string;
     name: string;
-    courseId: string; 
-    posts: string[]; 
+    courseId: string;
+    posts: string[];
 }
 
 export interface FollowupDiscussion {
     _id?: string;
-    postId: string; 
-    authorId: string; 
+    postId: string;
+    authorId: string;
     datePosted: string;
     content: string;
-    replies: string[]; 
+    replies: string[];
 }
 
 export interface Answer {
     _id?: string;
     type: number; // 0 = question, 1 = note
-    authors: string[]; 
+    authors: string[];
     content: string;
     dateEdited: string;
 }
 
 export interface Reply {
     _id?: string;
-    followupDiscussionId: string; 
-    authorId: string; 
+    followupDiscussionId: string;
+    authorId: string;
     datePosted: string;
     content: string;
 }
